Add documentation for most public items
This commit also directly increments all the necessary versions for
the next release.

As I was writing the documentation for the main traits + structs, I
realized that they actually expose more things publicly than needed.

Therefore, this commit also includes a breaking change for the tc_core
crate because some of the public methods are removed. For users of the
testcontainers crate, this shouldn't be noticable though since from the
PoV of testcontainers, it is a non-breaking change and thus a
patch-release.
