+++ core/src/lib.rs
@@ -11,7 +11,7 @@ mod docker;
 mod image;
 mod wait_for_message;
 
-pub use self::container::{Container, Logs, Ports};
-pub use self::docker::Docker;
+pub use self::container::Container;
+pub use self::docker::{Docker, Logs, Ports};
 pub use self::image::Image;
 pub use self::wait_for_message::{WaitError, WaitForMessage};
