+++ images/parity_parity/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tc_parity_parity"
-version = "0.2.0"
+version = "0.2.1"
 authors = [
     "Thomas Eizinger <thomas@coblox.tech>",
     "Philipp Hoenisch <philipp@coblox.tech>",
@@ -14,7 +14,7 @@ categories = ["development-tools::testing"]
 keywords = ["docker", "testcontainers", "parity", "ethereum"]
 
 [dependencies]
-tc_core = { path = "../../core", version = "0.1" }
+tc_core = { path = "../../core", version = "0.2" }
 log = "0.4"
 
 [dev-dependencies]
