--- testcontainers/.gitignore
@@ -1,5 +1,0 @@
-/target
-**/*.rs.bk
-Cargo.lock
-
-/.idea
